Moved RolePreserveService to XeniaDiscord.Common
